+++ src/zbdd.h
@@ -37,14 +37,14 @@ class SetNode : public NonTerminal {
   using NonTerminal::NonTerminal;  ///< Constructor with index and order.
 
   /// @returns Whatever count is stored in this node.
-  int count() const { return count_; }
+  long int count() const { return count_; }
 
   /// Stores numerical value for later retrieval.
   /// This is a helper functionality
   /// for counting the number of sets or nodes.
   ///
   /// @param[in] number  A number with a meaning for the caller.
-  void count(int number) { count_ = number; }
+  void count(long int number) { count_ = number; }
 
   /// @returns Cut sets found in the ZBDD represented by this node.
   const std::vector<std::vector<int>>& cut_sets() const { return cut_sets_; }
@@ -67,7 +67,7 @@ class SetNode : public NonTerminal {
 
  private:
   std::vector<std::vector<int>> cut_sets_;  ///< Cut sets of this node.
-  int count_ = 0;  ///< The number of cut sets, nodes, or anything else.
+  long int count_ = 0;  ///< The number of cut sets, nodes, or anything else.
 };
 
 /// @class Zbdd
@@ -161,9 +161,7 @@ class Zbdd {
   /// @returns The number of cut sets in ZBDD.
   ///
   /// @pre SetNode marks are clear (false).
-  ///
-  /// @warning Integer may overflow for a large ZBDD.
-  int CountCutSets(const VertexPtr& vertex) noexcept;
+  long int CountCutSets(const VertexPtr& vertex) noexcept;
 
   /// Cleans up non-terminal vertex marks
   /// by setting them to "false".
